Reorganized misc functions
